+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project (aims to) adhere to [Semantic Versioning](https://semver.org/sp
 ### Changed
 
 ### Fixed
+- themed busy cursor on native Wayland
 
 ---